pyproject.toml
[tool.poetry]
name = "ml_foundations_pytorch"
version = "0.0.1"
description = "A repo in which to investigate various pytorch architectures in calssification of MNIST and imagenet"
authors = ["Thomas O'Brien <thomas.obrien@zelim.co >"]
readme = "README.md"
packages = [
  {include = "ml_foundations_pytorch"}
]

[tool.poetry.dependencies]
python = "~3.11"
matplotlib = "*"
numpy = "~=1.23.5"
opencv-python = ">=4.1.1"
Pillow = "*"
PyYAML = "*"
requests = "*"
protobuf = "*"
encord = "*"
black = "*"
mypy = "*"
google-cloud-storage = ">=2.11.0"
wandb = "*"
imgaug = "*"
python-dotenv = "*"
pablo-common = {version = "^0.0.0", source = "pablo"}

[tool.poetry.group.onnx_export]
optional = false

[tool.poetry.group.onnx_export.dependencies]
onnx = ">=1.9.0"                                  # ONNX export
pybind11 = "*"
onnxsim = "*"
onnx-simplifier = ">=0.3.6" # ONNX simplifier
onnxruntime = ">=1.16.1"

[tool.poetry.group.torch.dependencies]
torch = "~=2.2.1"
torchvision = "*"

[tool.poetry.group.tf_export.dependencies]
tensorflow = ">=2.4.1"                            # TFLite export
tensorflow-io-gcs-filesystem = "*"
tensorflowjs = "*"                                # TF.js export
openvino-dev = "*"                                # OpenVINO export

[tool.poetry.group.jupyter.dependencies]
ipython = "*"               # interactive notebook
jupyter = "*"
ipykernel = "*"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff.lint.isort]
known_first_party = ["ml_foundations_pytorch"]
force_sort_within_sections = true

[[tool.poetry.source]]
name = "pablo"
url = "http://10.128.0.5:8081/repository/pypi-all/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
